Add 'unsafe-inline' to style sources in CSP
Unfortunately Google Maps won't work without this set:
https://issuetracker.google.com/issues/132600807

This means we'll need to disable it for now. This is still leaves us in
a better place than what we had before (no CSP at all), and is not as
dangerous an attack vector as unsafe inline scripts are.

This also requires us to apply the nonce *only* to the `script-src`
directive, as browsers will disregard 'unsafe-inline' for any directive
that includes a nonce.