(spike) Allow user to turn polygon from very precise to editable
With 1000s of points, a polygon is not editable unless you have hours to spare
to move the points.
With <20 points, the polygon displayed would not be accurate enough to the
area actually being searched.
So allow the user to reduce the number of points, when they click 'edit polygon'