Update vacancy-feedback-metrics-by-month.sql
- When calculating confidence intervals for various rates from hiring staff feedback on vacancies, take the population size into account. Previously we approximated confidence interval using the formula:

z * sqrt[ P * ( 1 - P ) / n ]

This assumed a large population size (N >> n). This assumption held much of the time. However a more accurate estimate is:

z * sqrt[ P * ( 1 - P ) / n ] * sqrt[ ( N - n ) / ( N - 1 ) ]

where P is the estimate of the probability that the thing being measured happens (e.g. the probably that a vacancy is a hire), z is a coefficient related to the confidence level (z=1.96 for a 95% confidence level), n is the sample size, and N is the population size.

- Remove legacy 'raw' metrics (i.e. unnormalised metrics which are less accurate than the normalised ones)