Refactor DSI API consumption when signing in
It's more reliable to check for the category to be e.g. LA than to
assume that an organisation is LA if it's neither trust or school,
or to assume that if the JSON contains a URN then it's a school.

This PR will remain in draft until I have confirmation from DSI that
the mappings from ids to organisation categories are correct.