Add timestamps to JobAlertFeedbacks table
Because records already exist in this table, we need to give them a
default created_at and updated_at.

I chose 16 October 2020 which is when we launched the job alert feedback
feature.

In this migration, after applying that default to all existing
records, we then reinstate the default 'null: false' restriction
which raises an error if you try to create a feedback without one of
these timestamps.