Set placeholder secret base for Webpacker compile
Now that we've added functionality to the app that relies on Rails's
secret key base, it needs to be present on any app/Rake run or the app
will refuse to boot.

This adds an inline placeholder `SECRET_KEY_BASE` environment variable
when the `webpacker:compile` step is run on Docker build. As this task
doesn't actually _use_ the secret key base, we can safely set it to an
arbitrary value.