Make Devise IP tracking columns strings
These have been generated as `inet` column types because the migration
ran using Postgres, which is normally fine as we run Postgres in
production too, but the BigQuery import library isn't able to deal with
these types. We don't really need any of the functionality Postgres
offers us to deal with IPs, so it's fine for these to be strings
instead.