Force signout of opposite user type on sign in
When signing in as a publisher or jobseeker, we need to sign out the
respective opposite user type to avoid getting the application into a
weird state or confuse our users. This is very much an edge case (we
don't expect either kind of user to have an account of the opposite
type), but needs to be covered nonetheless.

- Rename `Publishers::BaseController#check_user_last_activity_at` to
  `authenticate_publisher!` to be more consistent with Devise-style
  controller authentication enforcement
- Rename `Publishers::BaseController#update_user_last_activity_at` to
  to `update_publisher_last_activity_at`
- Add `Publishers::AuthenticationConcerns#sign_out_publisher!` to
  clear publisher-related session data (and call on jobseeker sign in)
- Add `Publishers::SignIn::BaseSessionsController::sign_out_jobseeker!`
  (and call on publisher sign in)
- Rename `sign_in_user` test helper to `sign_in_publisher`