Don't titlecase acronyms in SchoolGroup names
We receive names for Trusts from Get Information About Schools in a
state where all the letters are uppercase.

Exclude acronyms from titlecase transformation, so that for example
'GLF SCHOOLS' is stored as 'GLF Schools', not 'Glf Schools'.

Checking that there are no vowels will not catch all acronyms.
Hard-coded acronyms that the vowel-filter didn't catch.