Send page visited events to BigQuery
This is an initial version of sending events to BigQuery from within
our application (using the Streaming Insert API). It sends one event to
BigQuery for every (non-error) page view. In the future, we will want to
send other types of events as well, such as specific user actions or
errors, as well as including additional data, such as selected variants
for currently running AB tests.

- Create `Event` class to encapsulate an arbitrary event within the app
  (to be used in non-request lifecycle situations, e.g. within a
  background job)
- Create `RequestEvent` class to represent an `Event` that happens as
  part of the Rails request lifecycle (i.e. is triggered from a
  controller) and includes information about the context of the request
- Create `SendEventToDataWarehouseJob` to send event data to BigQuery
  asynchronously
- Trigger a basic `page_visited` event for every successful request
  (using `RequestEvent` with no additional metadata), except for PaaS
  healthchecks and API
- Set ActiveJob test queue adapter globally instead of on specific specs