Remove broken ActionDispatch trusted proxies
PR #269 introduced a change that set `action_dispatch.trusted_proxies`
to a list of AWS IP addresses on boot, ostensibly to deal with not
having access to the request's original client IP on the then-current
infrastructure setup (my guess would be that it was an ELB issue rather
than a Cloudfront issue as detailed in the original PR).

However, this ended up having the complete opposite effect on our now
current infrastructure: We have a private IP (in `10.0.0.0/8`) in the
`X_FORWARDED_FOR` header, which would normally be filtered out by
Rails's `ActionDispatch::RemoteIp` middleware. However, because setting
`trusted_proxies` _replaces_ the default trusted proxies list in its
entirety [1], we were no longer filtering out private IPs from the list
of forwarded-for IPs and therefore `request.remote_ip` would always
return a private IP.

- Remove custom `action_dispatch.trusted_proxies` setting from staging
  and production environment configuration
- Remove (now unnecessary) `AwsIpRanges` library and autoloading of
  `lib/modules`

[1] https://github.com/rails/rails/blob/master/actionpack/lib/action_dispatch/middleware/remote_ip.rb#L68