Add form validation errors to sign in page
Email format, email and password presence

Content has deemed it to be important for us to do some extension of Devise in order to make this very frequently used form more user-friendly.

I don't believe extending Devise in order to include a form model introduces too much code complexity
relative to the need for this page to be as user-friendlt as we can afford.

Devise requires us to use $resource_name ('jobseeker') for both the sign in and the sign up form.

This means the form labels can't be left implicit, since the two forms would point to the same translation namespace.