Append CF IP ranges to default trusted proxies
We need to filter out Cloudfront IP ranges _in addition_ to private IP
ranges from our load balancers. The way the `RemoteIp` middleware
works, assigning the Cloudfront IPs to `action_dispatch.trusted_proxies`
_replaces_ the existing set of trusted proxies (which includes private
IP ranges) instead of appending to it, so we need to merge them with
the default set.