Add map to visualize/debug polygons
For business analysts and others to visualize the polygons searched in.

To see a map of the polygon(s) searched in, just add '+map' to the
end of the search query in the location field. E.g.:

Keywords
[english]
Location
[Hampshire+map]

This is for internal use only. It's password-protected for that reason,
but we do not need to maximize security here as there is no data at
risk.

I am using HTTP basic auth for the password protection.

The password and username are environment variables which can be found
in the usual store of environment variables.