[TEVA-1763] Add turnstyle anti-concurrency waiting to deploy_app.yml
The "calling" workflow deploy.yml uses the turnstyle action to ensure
only a single copy of the workflow is running at any one time. It then
uses workflow dispatch to start the deploy_app workflow, for staging,
then production.
To cope with multiple merges in quick succession, introduce queuing with
turnstyle in the "called" workflow.