Enable searching in certain locations by using component districts
These are Bedfordshire, Berkshire, Cheshire, Greater Manchester, Merseyside,
Tyne and Wear, West Yorkshire, Yorkshire. Most of these are counties.

These locations are often searched, but they are not included in ONS' API for county
and unitary authority boundaries, which we use to import most county boundaries.
That API concerns local government divisions, rather than what people commonly think of counties,
and so excludes what are known as 'ceremonial' or 'geographical' counties:
https://en.wikipedia.org/wiki/Ceremonial_counties_of_England

As of January 2021 I could not find a usable source of ceremonial county boundaries, so I decided
to construct these out of their component local authorities. I am also including 'Yorkshire',
which is not a ceremonial county, because it's a common search location.

Also in this commit:

- Import all unitary authorities
- Do not distinguish between London boroughs and counties/unitary authorities. These all come
from the same API endpoint.
- Omit cities that have the same name as a unitray authority so that we search in the larger
area.
- Do not fall back to a text-based search if the polygon is missing; use the point instead.
(This is necessitated by us no longer distinguishing between unitary authorities and counties.)
- Reduce number of API calls to buffer service:
When there is no change in the boundary, we don't need to ask ONS
to recalculate the buffers around it.