Refactor publisher vacancies sorting options
We previously assigned translations to a constant on app boot time,
which does not reliably work (as the I18n translations may not have been
loaded yet). Instead, this is extracted to a `VacancySortingOptions`
object which loads translations at runtime (and acts as an `Enumerable`
so can be directly used by the `collection_select` view helper).