Update to use latest ONS API endpoint
We were directed by ONS to use the https://services1.arcgis.com
API instead of https://ons-inspire.esriuk.com when we reported
a bug with ons-inspire (the old API).

Updating the API will solve some issues peculiar to the old API, such
as Liverpool being in Wales.

In the process of updating this I found some mistakes we had made when
we first implemented polygon importing. We previously took the first ring
from each set of rings. This is not always the right ring to use: there are
a dozen or so exceptions.

When the ONS API returns multiple rings for a county, some of these
are small islands, one is the mainland of the county, and some are cities
that are excluded because they are locally governed.

There is no pattern as to which of the rings is relevant to us, so we have
to individually specify which ring is the one we want
(generally the largest), using 0-indexing.

The way we import polygons is not sophisticated enough to include all
schools for each county. The schools which we are failing to capture are:
- In Essex, which has 10 rings, the 5th (I am not 0-indexing here) is
Mersea Island, containing Mersea Island School.
- In Hampshire, which has 2 rings, the 1st is Hayling Island,
containing at least one school.
- In Kent, which has 4 rings, the 3rd is Sheppey Island,
containing at least one school.

The Isles of Scilly only has one school at the time of writing,
so we use the ring for St Mary's Island:
https://en.wikipedia.org/wiki/List_of_schools_on_the_Isles_of_Scilly

This is an interim solution to updating the API, until we decide how
best to deal with the omitted island schools.
TODO: Deal with the omitted island schools. This will be easier when
we are doing requests with multiple polygons e.g. for composite locations.

URL_MAXIMUM_LENGTH adjusted to 33_000 after testing. The endpoint can cope
with a length of 33742, but was broken by a call with length 105_753.