Update import task to assign LocationPolygon location_type differently
Our app needs two different senses of 'county' and 'city', one for ONS, and one for humans. The human-friendly one is needed for the vacancy facets on the front page. The ONS one is needed to define which locations in the API response are in scope. ONS' API endpoints don't separate places in the same way the English language does, but using their own definitions.

This commit adds to the import task a mapping from the ONS categorisation to regular
human categorisation; the latter is what we store against 'location_type' on the
LocationPolygon in the database (these types are not used very much at the moment.)