Use multiple polygons ('rings') per LocationPolygon
This step (4) completes the code changes needed to consume the new
data structure for LocationPolygon.

See here for the plan and context:
https://github.com/DFE-Digital/teaching-vacancies/pull/2593#issuecomment-764615951

The biggest change in the commit not already explained in PR/commit
message for Step 2 https://github.com/DFE-Digital/teaching-vacancies/pull/2706:

Enable searching in certain locations by using component districts

These are Bedfordshire, Berkshire, Cheshire, Greater Manchester, Merseyside,
Tyne and Wear, West Yorkshire, Yorkshire. Most of these are counties.

These locations are often searched, but they are not included in ONS' API for county
and unitary authority boundaries, which we use to import most county boundaries.
That API concerns local government divisions, rather than what people commonly think of counties,
and so excludes what are known as 'ceremonial' or 'geographical' counties:
https://en.wikipedia.org/wiki/Ceremonial_counties_of_England

As of January 2021 I could not find a usable source of ceremonial county boundaries, so I decided
to construct these out of their component local authorities. I am also including 'Yorkshire',
which is not a ceremonial county, because it's a common search location.

- Do not fall back to a text-based search if the polygon is missing; use the point instead.
(This is necessitated by us no longer distinguishing between unitary authorities and counties.)
- Map 'manchester' to 'greater manchester'
The Manchester unitary authority shape doesn't really match
what people mean when they say 'Manchester'.
- Rename location_category_search. We will soon move away from having the location_category param.
- Remove LocationCategory. This is a left-over class which predated the use of polygons.
This behaviour is related to polygons and the use of 'category' as a
term is both confusing and being phased out.

This commit also fixes a bug/adds the feature that mapped_locations
can also work for composite locations.