Consolidate 5 existing feedback tables into one
This new consolidated table is not yet used by the code, so the
task will have to be run after that code is updated.

I tested the task on production data locally and checked
feedbacks of all types.

- Migration for new feedback table
- Task for one-off transfer of data from old tables
- Do not export feedbacks table because we will send data as
events
- There is a sixth feedback type coming up, for application
submission
- Drop unused data in ratings columns on GeneralFeedback
and VacancyPublishFeedback - we stopped adding to these
in 2019
- Rename not_interested scopes to uninterested to avoid
warning messages about scope conflict whenever you run
the console
- Rename some other columns to be specific
- Consolidate various 1200-char columns into a single column
called 'comment'
- Remove unused scope