Fix for NoMethodError when location is nil
Due to the problems caused by the remove_location_category_from_search_criteria task, location was set to nil on many subscriptions. As location was nil, when .empty? is called on location, a NoMethodError is raised. Now this has been changed to .blank?, line 48 will evaluate to true if the location is nil.