Send Events to BigQuery at the same time as moving feedback data
This updates the task which sends feedback from old feedback tables to the
new feedback table, so that we are creating Events (sent to BigQuery) at
the same time. These events will be used alongside newer RequestEvents,
representing controller events that create Feedbacks, to store/manage/
monitor/analyse/collate all feedback of a certain 'feedback_type'.

The task has not yet been run. It will be run when all code has been
updated to use the new Feedback table rather than the old feedback table.

I'm using find_or_create_by in case GovUK PaaS explodes and we need to
run the task a second time.

I have also updated the Event model so that any data containing a
'created_at' attribute is reflected in the 'occurred_at' column of the
Event (rather than Time.now); and so that fewer data value types are
converted to string (which can be an annoyance for someone wanting to
manipulate the data after it's received).