Add clause to task to prevent corruption of data
Previous to these changes, the task did not check that location_category was present and location was blank before setting location to the value of location_category and deleting location_category. Because of this, all of the subscriptions that did not have a location_category in the search_criteria set location to nil - the return value of subscription.search_criteria.delete[location_category] is nil if location_category is not present.