Add subscriptions incident recovery Rake task
The last piece of the puzzle to recovering from the subscription
search criteria incident. This Rake task loads a JSON export of
subscription creation events, correlates them with records created
after the last good backup (subscriptions before that have already
been fixed) but before the incident (subscriptions created after
are not affected).

It requires a parameter of a local file (to be created on the
instance where the task is run) containing the event data:

```
rails location_fix[data.json]
```

Once the changes have been verified with the default dry run,
there is a second "commit" parameter to update subscriptions:

```
rails location_fix[data.json,true]
```