Update Job Alert Feedback controller/form to rely on new Feedback table
And use trigger_feedback_provided_event to create Events. This method uses
feedback_params, resulting in a slightly unconventional feedback_params
method on JobAlertFeedbacksController, since this is the only type of
feedback that has more than one action (they can be updated as well as
created).

Also: update Event#format_value to keep floats and integers rather than
transform them to strings. The idea here is to save deconverting them
after they are sent to BigQuery.

This includes a single test for links in job alert emails already sent,
in order to avoid a bug such as that fixed in
https://github.com/DFE-Digital/teaching-vacancies/pull/2737.