Include search criteria info in unsubscribe/job alert feedback
This info is handy to include, partly because the search criteria may be edited
after the feedback is submitted, and partly for convenience.

Added matcher to check two json strings match which are not necessarily in the same order,
since hashes can vary their orders. Named by analogy with existing hash_including method.