Start reporting CSP violations again
We had turned this off because the Rollbar gem generated a CSP violation
on every pageview due to it skipping the CSP nonce if `unsafe_inline`
was set (which we do for backwards compatibility), but this has been
fixed in 3.1.2 and there seem to be no spurious CSP violations left.

We'll continue running in report-only mode for the time being, but in
theory we should no longer receive any reports now!