Allow notification component to customize html attributes
This is so that the 'notice'-styled version of the component can be used for the info panel
on the application submission confirmation page.

By default, notification components will render with the 'alert' role and tabindex -1, which
is not the correct role/tabindex for the info panel on this page (affects screen
readers).