Make notification components less hostile to screen-readers
The issue which the html_attributes attribute tries to solve is when we
reuse the notification component as something other than a notification,
e.g. for an 'empty' box (such as 'You have no saved jobs'), or for an
info box (such as 'You can unsubscribe in such and such a way'), we don't
want to give it the html role of 'alert' or the tabindex of '-1' which
cause screen-readers to get to the non-notification before anything else
and mess with the tab-order.

This can be refactored in the very near future but I wanted to get this
piece of accessibility in first so that it can be retained when the
refactoring happens - but maybe doing things in the other order
(refactor than accessibilize) is better?