Clamp Algolia expiry filter to start of hour
This is the least effort way of making Algolia queries a bit easier to
cache. Because we need to filter on `expires_at_timestamp` at query
time, we can't cache any given query because the next one will have a
different timestamp.

This means we may briefly include expired results in the search results
for up to 59 minutes.