Limit API to only return live vacancies
Our API has a fair share of issues, and we need to take a step back and
rethink it in the future.

Until then, we need to limit vacancies being returned to just be live
vacancies as existing users are crawling hundreds of pages of expired
vacancies which is causing issues, and potential new users don't want
to deal with this overload.