Update text shown while browser finds user location
This text is only displayed for long enough to read it if the browser is taking longer than expected to find the location. permissions This probably means that the user has disabled location services in some settings (system-wide, browser, etc.) If location services are disabled, the current behaviour is that the wheel keeps spinning indefinitely and the text, misleadingly, says "finding location". If the text says "finding location", the user may take that to mean that the computer is performing the task correctly; "requesting" is more likely to hint to the user that the reason it has stalled is because of permissions.