Only use Rails subgems we actually need
This came out of a subgem of Rails (`activestorage`) using a dependency
that was yanked, blocking us from bundling our dependencies. It should
make our app leaner and reduce the surface area of these sorts of things
happening a tiny bit.

We still need to do something about `govuk-date-fields` because that is
still pulling in _all_ of Rails, but this is a first step.

- Split out individual gems in Gemfile
- Update `govuk-components` to new version that doesn't depend on all of
  Rails
- Update `mimemagic` to new version
- Update Dockerfile to install `shared-mime-info` (new `mimemagic`
  runtime dependency)