[TEVA-2087] Destroy review only after deploy completes
When a PR is closed soon after it has been created, it leads to a terraform
race condition between destroy.yml and review.yml.
destroy.yml is likely to complete first, but runs against non-existent
infrastructure because it is still being spun up.
By the time deploy_app.yml spins up the infrastructure, there is no
automated way of destroying the infrastructure subsequently.

To simplify the dependencies between jobs, we now generate a unique tag
which can be used in both trigger and wait actions in a consistent way.

We also simplify the docker tag by removing the unnecessary timestamp so
the job name is predictable.

The timeout on deploy has been increased to 20min as it timed out during
testing.