Consolidate policies for deployment role
We've run into an AWS limit with the number of policies we can attach
to a role (10). This consolidates all the policies that are unique to
the deployments role into a single policy.

This is not as reusable as having individual policies, but fixes the
limit issue. We can always split parts out of it later if we need to.