Refactor qualifications step
Our treatment of secondary qualifications as individual qualifications
that are only summarised at display level has proven a bit of a struggle
when it came to allowing users to add/edit them. This introduces a
sub-level of "qualification results" (of which a qualification can have
many), and finishes implementation of the qualifications step for
secondary qualifications (pending some extra tests and refactoring).

- Create `QualificationResult` as an associated model of `Qualification`
  to represent subjects/grades for secondary qualifications
- Add `CommonForm#qualification_results_attributes=` to allow
  use of `fields_for` in the views
- Make use of `accepts_nested_attributes_for` to allow `Qualification`
  to handle persisting nested `QualificationResult`s once the form
  object validations have passed
- Add handling of qualification results in "Quick Apply"
- Remove handling of "grouped" qualifications everywhere (now that each
  qualification is a top-level entity)