Send equal opportunities report publication events to BigQuery
As this data is sensitive, and could be deduced from matching up timestamps of the updated equal_opportunities_reports and job_applications, we will need to handle it separately and differently than we do normal entity create/update/destroy events, rather than sending equal_opportunities_reports to BigQuery whenever the report is updated (which is the same thing as 'whenever a job application is submitted').

A new job runs every night after midnight. This works by triggering an event for every job listing that expired yesterday, but only if the listing has more than [threshold] applications, to preserve anonymity in the streamed events db.