Remove hiring staff access to applications after 1 year
This implements the minimal code required for the feature, because
we will not need the feature until a year from now.

If we want to at the time, we can spruce it up and, for example,
tell the user why they can't view the application.

But for now we can save time by doing the minimal change that makes
us compliant.

Another future approach could be to remove the 'view application'
link.