Remove state column from vacancies.
This field initially arose from a performance analysis need, and ended up being used for things it shouldn't have.
This need no longer exists, so we should remove the field.
This commit does that, but prioritises maintaining functionality over code quality.

This commit is part of a wider refactor around the create a job journey.