Refactor parts of the create a job journey.
- Remove vacancy wizard back link component
- Add vacancy helper method to base controller
- Add form helper method to build controller
- Only pass the attributes for a given step to the form model
- Pass the actual vacancy to the form model rather than initializing a new vacancy that is mostly blank

This commit is part of a broader refactor that will help refactor the important dates step, and also
will help bring the 'Create a job' journey more in line with the 'Apply for a job' journey in terms
of code style, design patterns, testing etc. This reflects ongoing design work that will also seek to
make things more consistent across the application as a whole.