Fix delete review app workflow
We were waiting on the deploy-app job on the wrong commit. In case of a
PR, deploy-app is linked to the merged commit, whereas we were looking
as the PR branch HEAD commit.

Now we wait on the calling job "Deploy app", in the "review" matrix
iteration. It's also based on the PR branch HEAD commit.