Ensure deleting documents also removes supporting docs
While we are running the legacy Google Drive documents and the new
ActiveStorage backed supporting documents in parallel, we need to make
sure we actually delete supporting documents when their corresponding
document gets deleted.

This is not the nicest way of doing it, but it gets the job done, and
will be gone once we remove all Google Drive handling.