Anonymise vacancy ids in request events
In order to match up the events with records, we need to use the anonymised version.

- Allow click_events to have further event data (from a click_event_data param)
to include in the event. The use case here is to match up a 'save vacancy' click
event with the vacancy in question. We have to explicitly permit such parameters
in order to convert the click_event_data param into a hash for consumption by
Event.