Add a fake DSI sign out endpoint for testing
This adds a `FakeDsiSignOutEndpoint` Rack middleware that intercepts
requests to a fake URL that it overrides `DFE_SIGN_IN_ISSUER` with.

This simulates the proper flow a user would go through for the
purpose of system tests checking that hiring staff can sign out
properly, and fixes an issue where tests will fail locally if a
value is set for `DFE_SIGN_IN_ISSUER`.