Refactor the conversion of Devise flash variants
Previously, we used methods such as 'replace_devise_notice_flash_with_success' and
'replace_devise_alert_flash_with_notice' on particular actions in order to
convert flash messages originating from Devise into the variant we wanted
(alert, notice, success variants).

This was unexpected, from the point of view of Devise's internals, and resulted
in odd behaviour such as flash messages persisting for one extra page load. Devise
expected certain messages to be stored in certain flash variants.

Therefore, instead of manipulating the variant of the flash itself, I have added
some logic that alters the 'variant' parameter passed to FlashComponent, mapping
all Devise alert flashes to pass 'notice' as the variant to FlashComponent, and
mapping 'notice' to 'success'. This preserves the same variants as we were already
using. Any messages that are not from Devise are not altered.